Create module with common plotting and data io code
