+++ cefi_kit/src/cefi_kit/__init__.py
@@ -1,0 +1,2 @@
+def hello() -> str:
+    return "Hello from cefi-kit!"
