+++ diagnostics/biogeochemistry/zooplankton.py
@@ -18,7 +18,9 @@ import sys
 # Get the directory of the current script
 script_dir = os.path.dirname(os.path.abspath(__file__))
 sys.path.append(os.path.join(script_dir, '../physics'))
-from plot_common import add_ticks, corners, open_var
+from cefi_kit.io import open_var
+from cefi_kit.plots import add_ticks
+from plot_common import corners
 
 PC = ccrs.PlateCarree()
 
@@ -31,7 +33,7 @@ def plot_zooplankton(pp_root, label):
         season_data = season_data['Total Carbon Mass (mg-C/m3)'].to_xarray()
         season_data['season'] = s
         seasonal_obs.append(season_data)
-        
+
     seasonal_obs = xarray.concat(seasonal_obs, dim='season')
     seasonal_obs = seasonal_obs.rename({'Latitude': 'lat', 'Longitude': 'lon'})
     seasonal_obs *= 2.0 # correction for underestimation
@@ -55,35 +57,35 @@ def plot_zooplankton(pp_root, label):
         # mom6
         for row in [0, 2]:
             p = axs[row, col].pcolormesh(
-                model_grid.geolon_c, 
-                model_grid.geolat_c, 
-                mesozoo_vol.sel(season=s).squeeze(), 
+                model_grid.geolon_c,
+                model_grid.geolat_c,
+                mesozoo_vol.sel(season=s).squeeze(),
                 **common
             )
             i = row*4 + col
             axs[row, col].set_title(f'({ascii_lowercase[i]}) {s} model', fontsize=11)
-        
+
         # copepod
         for row in [1, 3]:
             axs[row, col].pcolormesh(xcorner, ycorner, seasonal_obs.sel(season=s), **common)
             i = row*4 + col
             axs[row, col].set_title(f'({ascii_lowercase[i]}) {s} COPEPOD', fontsize=11)
             axs[row, col].coastlines('50m')
-    
+
         yl = 1 if s == 'SON' else 0
-        # top 2 
+        # top 2
         for row in [0, 1]:
             add_ticks(axs[row, col], xticks=np.arange(-100, -31, 5), yticks=np.arange(0, 61, 5), xlabelinterval=1, ylabelinterval=yl, fontsize=7)
             axs[row, col].set_xlim(-82, -59)
             axs[row, col].set_ylim(25, 46)
-        
+
         # bottom 2
         for row in [2, 3]:
             add_ticks(axs[row, col], xticks=np.arange(-100, -31, 5), yticks=np.arange(0, 61, 5), xlabelinterval=1, ylabelinterval=yl, fontsize=7)
             axs[row, col].set_xlim(-98.5, -79.5)
             axs[row, col].set_ylim(18, 33)
-        
-    
+
+
     f.subplots_adjust(bottom=0.23, wspace=0.1, hspace=.3)
     cbax = f.add_axes([0.3, 0.15, 0.4, 0.02])
     cb = f.colorbar(p, cax=cbax, orientation='horizontal', extend='max')
@@ -98,7 +100,7 @@ def plot_zooplankton(pp_root, label):
         for s in ax.spines.values():
             s.set_visible(False)
 
-    if label == '':            
+    if label == '':
         plt.savefig('figures/mesozoo.png', dpi=200, bbox_inches='tight')
     else:
         plt.savefig('figures/mesozoo.png', dpi=200, bbox_inches='tight')
