+++ diagnostics/physics/NWA12/tbot_epu.py
@@ -2,7 +2,7 @@
 Compare annual average bottom-temperature anomalies in four different northeast US ecological production units from model,
 reanalysis, and observed data
 How to use:
-python tbot_epu.py /archive/acr/fre/NWA/2023_04/NWA12_COBALT_2023_04_kpo4-coastatten-physics/gfdl.ncrc5-intel22-prod/ 
+python tbot_epu.py /archive/acr/fre/NWA/2023_04/NWA12_COBALT_2023_04_kpo4-coastatten-physics/gfdl.ncrc5-intel22-prod/
 """
 
 from datetime import datetime, timedelta
@@ -10,13 +10,8 @@ import matplotlib.pyplot as plt
 import pandas as pd
 from string import ascii_lowercase
 import xarray
-import os
-import sys
 
-# Get the directory of the current script
-script_dir = os.path.dirname(os.path.abspath(__file__))
-sys.path.append(os.path.join(script_dir, '../'))
-from plot_common import open_var
+from cefi_kit.io import open_var
 
 
 # Convert decimal date to a datetime object
@@ -42,7 +37,7 @@ def load_ecodata(epu):
 
 def epu_average(var, area):
     ave = var.weighted(area).mean(['yh', 'xh']).compute()
-    # Note anomalies from ecodata/SOE are supposed to be relative to a 1981--2010 climatology. 
+    # Note anomalies from ecodata/SOE are supposed to be relative to a 1981--2010 climatology.
     anom = ave.groupby('time.month') - ave.sel(time=slice('1993', '2010')).groupby('time.month').mean('time')
     # Resample so data is indexed by start of month
     monthly = anom.resample(time='1MS').first()
@@ -75,15 +70,15 @@ def plot_tbot_epu(pp_root):
         glo = glorys['tob_anom'].sel(EPU=epu).to_series()
         obs = load_ecodata(epu)
         combined = pd.concat((mod, glo, obs), axis=1, keys=['mod', 'glorys', 'obs']).dropna()
-        annual = combined.resample('1AS').mean()
+        annual = combined.resample('1YS').mean()
         obs_corr = annual.corr().loc['mod', 'obs']
         glo_corr = annual.corr().loc['mod', 'glorys']
         annual['mod'].plot(label='Model', c=model_color, lw=lw, ax=ax)
-        annual['glorys'].plot(label=f'GLORYS12 (r={glo_corr:2.2f})', c=glorys_color, lw=lw, ax=ax)    
-        annual['obs'].plot(label=f'Observed (r={obs_corr:2.2f})', c='k', lw=lw, ax=ax)  
+        annual['glorys'].plot(label=f'GLORYS12 (r={glo_corr:2.2f})', c=glorys_color, lw=lw, ax=ax)
+        annual['obs'].plot(label=f'Observed (r={obs_corr:2.2f})', c='k', lw=lw, ax=ax)
         name = long_names_ordered[epu]
         ax.set_title(f'({letter}) {name}')
-        
+
     for ax in axs.flat:
         ax.set_xlim('1992', '2020')
         ax.set_xlabel('')
