+++ pyproject.toml
@@ -1,0 +1,46 @@
+[project]
+name = "cefi-kit"
+version = "0.1.0"
+description = "Add your description here"
+readme = "README.md"
+authors = [
+    { name = "Andrew C. Ross", email = "andrew.c.ross@noaa.gov" }
+]
+requires-python = ">=3.13,<3.14"
+dependencies = [
+    "esmpy>=8.8.1",
+    "gsw>=3.6.23",
+    "h5py>=3.16.0",
+    "netcdf4>=1.7.4",
+    "numpy>=2.4.6",
+    "pandas<3.0", # < 3 needed for read_excel date_parser in hypoxic area
+    "xarray>=2026.7.0",
+    "xesmf>=0.9.2",
+    "xskillscore>=0.0.29",
+]
+
+[project.optional-dependencies]
+diagnostics = [
+    "cartopy>=0.25.0",
+    "colorcet>=3.2.1",
+    "cmcrameri>=1.10",
+    "matplotlib<3.11", # need to preserve plt.cm.get_cmap for now
+    "openpyxl>=3.1.5",
+    "utide>=0.3.1",
+]
+
+[build-system]
+requires = ["uv_build>=0.11.14,<0.12.0"]
+build-backend = "uv_build"
+
+# On ppan, the initial uv sync should happen with the module python/3.13 loaded.
+# After that, unload python, source .venv/bin/activate[.csh], module load esmf.
+# This is necessary for xesmf and esmpy to load the right libs and work correctly.
+[tool.uv]
+python-preference = "system"
+
+[tool.uv.build-backend]
+module-root = "cefi_kit/src"
+
+[tool.uv.sources]
+esmpy = { git = "https://github.com/esmf-org/esmf", subdirectory = "src/addon/esmpy", tag = "v8.8.1" }
